postgres creation scripts
One script creates the schema and the tables, the other takes the previously generated schema and tables and builds the first script.

This allows me to continue to add records directly to the tables in the database and then generate a script.

NOTE: No validation or error-checking of any kind is performed.